Change banded_sw to faster and less memory consuming solution Add composition bias flag remove some warnings

+++ src/prefiltering/SubstitutionMatrixWithoutX.h
@@ -11,8 +11,8 @@ class SubstitutionMatrixWithoutX : public BaseMatrix {
     public:
     SubstitutionMatrixWithoutX(double **probMatrix, float **rMatrix, short **subMat, float bitFactor) {
         // copy data
-        for(size_t i = 0; i < this->alphabetSize; i++) {
-            for (size_t j = 0; j < this->alphabetSize; j++) {
+        for(int i = 0; i < this->alphabetSize; i++) {
+            for (int j = 0; j < this->alphabetSize; j++) {
                 this->probMatrix[i][j] = probMatrix[i][j];
             }
         }
@@ -21,8 +21,8 @@ class SubstitutionMatrixWithoutX : public BaseMatrix {
 //                this->subMatrixPseudoCounts[i][j] = rMatrix[i][j];
 //            }
 //        }
-        for(size_t i = 0; i < this->alphabetSize; i++) {
-            for (size_t j = 0; j < this->alphabetSize; j++) {
+        for(int i = 0; i < this->alphabetSize; i++) {
+            for (int j = 0; j < this->alphabetSize; j++) {
                 this->subMatrix[i][j] = subMat[i][j];
             }
         }
